+++ upyog/imports.py
@@ -3,6 +3,7 @@ import enum
 import functools
 import inspect
 import json
+import math
 import mimetypes
 import operator
 import os