+++ upyog/utils/utils.py
@@ -3,7 +3,7 @@ from itertools import islice
 
 __all__ = [
     "flatten", "uniqueify", "get_YYYY_MM_DD", "chunk", "notnone",
-    "lmap", "allequal", "zipsafe"
+    "lmap", "allequal", "zipsafe", "convert_size"
 ]
 
 
@@ -51,3 +51,15 @@ def zipsafe(*its):
     if not allequal(lmap(len, its)):
         raise ValueError("The elements have different leghts")
     return zip(*its)
+
+
+# https://stackoverflow.com/questions/5194057/better-way-to-convert-file-sizes-in-python
+def convert_size(size_bytes: int) -> str:
+    "Convert bytes into appropriate unit and return a human readable string for file size"
+    if size_bytes == 0:
+       return "0B"
+    size_name = ("B", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB")
+    i = int(math.floor(math.log(size_bytes, 1024)))
+    p = math.pow(1024, i)
+    s = round(size_bytes / p, 2)
+    return "%s %s" % (s, size_name[i])